Correct win64 fix
Part of #69

+++ src/prodbg/ui/ui_dock.cpp
@@ -975,8 +975,8 @@ void UIDock_updateSize(UIDockingGrid* grid, int width, int height)
 {
 	Vec2 deltaMove; 
 
-	deltaMove.x = (int)(width - grid->rect.width);
-	deltaMove.y = (int)(height - grid->rect.height);
+	deltaMove.x = (float)(width - grid->rect.width);
+	deltaMove.y = (float)(height - grid->rect.height);
 
 	grid->rect.width = width;
 	grid->rect.height = height;
